+++ backend/config/database.js
@@ -1,0 +1,17 @@
+const mongoose = require('mongoose');
+require('dotenv').config({ path: './config/.env' });
+
+const connectDB = async () => {
+  try {
+    await mongoose.connect(process.env.DB_URL, {
+      useNewUrlParser: true,
+      useUnifiedTopology: true
+    });
+    console.log('MongoDB connection successful');
+  } catch (err) {
+    console.error('MongoDB connection failed:', err);
+    process.exit(1); // Exit process with failure
+  }
+};
+
+module.exports = connectDB;
