--- frontend/src/App.jsx
@@ -1,14 +1,0 @@
-// src/App.jsx
-
-import { BrowserRouter, Routes, Route } from "react-router-dom";
-import Home from "./pages/Home";
-
-const App = () => (
-  <BrowserRouter>
-    <Routes>
-      <Route path="/" element={<Home />} />
-    </Routes>
-  </BrowserRouter>
-);
-
-export default App;
