--- frontend/src/components/Navbar.jsx
@@ -1,29 +1,0 @@
-//src/components/Navbar.jsx
-
-import { Menu } from "lucide-react";
-import { Button } from "./ui/button";
-import { Avatar, AvatarFallback, AvatarImage } from "./ui/avatar";
-
-const Navbar = ({ toggleSidebar, isSidebarOpen }) => {
-  return (
-    <header className="sticky top-0 z-30 w-full border-b bg-background">
-      <div className="flex h-16 items-center px-4 w-full">
-        <div className="flex items-center">
-          <Button variant="ghost" size="icon" onClick={toggleSidebar} className="mr-2">
-            <Menu className="h-5 w-5" />
-            <span className="sr-only">Toggle sidebar</span>
-          </Button>
-          <h1 className="text-xl font-semibold">NoteNest</h1>
-        </div>
-        <div className="ml-auto flex items-center">
-          <Avatar>
-            <AvatarImage src="https://github.com/shadcn.png" />
-            <AvatarFallback>UN</AvatarFallback>
-          </Avatar>
-        </div>
-      </div>
-    </header>
-  );
-};
-
-export default Navbar;
