--- frontend/src/components/ui/button.jsx
@@ -1,34 +1,0 @@
-import React from "react";
-import clsx from "clsx"; // optional, for class merging
-
-export const Button = ({
-  children,
-  variant = "default",
-  size = "default",
-  className = "",
-  ...props
-}) => {
-  const baseStyles =
-    "inline-flex items-center justify-center rounded-md font-medium transition-colors focus:outline-none disabled:opacity-50 disabled:pointer-events-none";
-
-  const variants = {
-    default: "bg-blue-600 text-white hover:bg-blue-700",
-    ghost: "bg-transparent hover:bg-gray-100 dark:hover:bg-gray-300",
-  };
-
-  const sizes = {
-    default: "h-10 px-4 py-2",
-    icon: "h-10 w-10 p-2",
-  };
-
-  return (
-    <button
-      className={clsx(baseStyles, variants[variant], sizes[size], className)}
-      {...props}
-    >
-      {children}
-    </button>
-  );
-};
-
-export default Button;
